LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/nested-non-auto-inline-size-offset-top-expected.txt


FAIL offsetTop on element in second outer column assert_equals: expected 0 but got 90
